# Beispielanwendung Adressbuch/Backend/src/api/openapi.yaml
openapi: 3.0.0
info:
  title: Adressbuch
  description: "Die Adressbuch-App nutzt einen simplen REST-Webservices zur Kommunikation\
    \ mit dem Backend. Dieser ermöglicht die üblichen CRUD-Operationen zur Suche,\
    \ Anzeige, Bearbeitung und Löschung von Adressen, wobei jede Adresse aus einem\
    \ Vornamen, Nachnamen, Telefonnummer und eine E-Mailadresse besteht."
  contact:
    name: Dennis Schulmeister-Zimolong
    url: https://www.wpvs.de
    email: dhbw@windows3.de
  license:
    name: CC-BY-4.0
    url: https://creativecommons.org/licenses/by/4.0/deed.de
  version: "1.0"
servers:
- url: http://localhost:3000
tags:
- name: Index
- name: Adresse
paths:
  /:
    get:
      tags:
      - Index
      summary: Wurzeladresse des Webservices
      description: Liefert eine HATEOAS-Übersicht aller collections
      operationId: index
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HATEOAS_Collection'
                x-content-type: application/json

  /address:
    get:
      tags:
      - Adresse
      summary: Adressen suchen
      description: Suche nach Adressen. Gibt eine Liste aller Adressen passend zu
        den übergebenen Suchkriterien zurück.
      operationId: search-address
      parameters:
      - name: first_name
        in: query
        description: Vorname
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: last_name
        in: query
        description: Nachname
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Address'
                x-content-type: application/json
    post:
      tags:
      - Adresse
      summary: Adresse anlegen
      description: "Neue Adresse anlegen. Speichert die übergebene Adresse in der\
        \ Datenbank. Der Wert des Feldes \"_id\" wird dabei ignoriert, da in jedem Fall\
        \ eine neue Adresse mit neuer ID angelegt wird."
      operationId: create-address
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Address'
      responses:
        "201":
          description: Created
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Address'

  /address/{id}:
    get:
      tags:
      - Adresse
      summary: Adresse lesen
      description: Liest eine einzelne Adresse vom Server.
      operationId: get-address
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Address'
        "404":
          description: Not Found
    put:
      tags:
      - Adresse
      summary: ""
      description: "Überschreibt eine vorhandene Adresse auf dem Server. Bis auf die\
        \ ID können alle Felder geändert werden. Der Versuch, die ID zu ändern, wird\
        \ vom Server stillschweigend ignoriert."
      operationId: update-address
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Address'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Address'
        "404":
          description: Not Found
    delete:
      tags:
      - Adresse
      summary: Adresse löschen
      description: Löscht eine Adresse auf dem Server.
      operationId: delete-address
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
    patch:
      tags:
      - Adresse
      summary: Einzelne Adressfelder ändern
      description: "Ändert einzelne Felder einer vorhandenen Adresse. Zulässig sind\
        \ alle Felder bis auf die ID, da diese nachträglich nicht mehr verändert werden\
        \ darf."
      operationId: patch-address
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Address'
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
components:
  schemas:
    HATEOAS_Collection:
      title: HATEOAS-Links einer Collection
      type: object
      properties:
        _name:
          type: string
        query:
          $ref: '#/components/schemas/HATEOAS_Link'
        create:
          $ref: '#/components/schemas/HATEOAS_Link'

    HATEOAS_Resource:
      title: HATEOAS-Links einer Resource
      type: object
      properties:
        read:
          $ref: '#/components/schemas/HATEOAS_Link'
        update:
          $ref: '#/components/schemas/HATEOAS_Link'
        patch:
          $ref: '#/components/schemas/HATEOAS_Link'
        delete:
          $ref: '#/components/schemas/HATEOAS_Link'

    HATEOAS_Link:
      title: HATEOAS-Link
      type: object
      properties:
        url:
          type: string
        method:
          type: string

    Address:
      title: Adresse
      type: object
      properties:
        _links:
          $ref: '#/components/schemas/HATEOAS_Resource'
        _id:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        phone:
          type: string
        email:
          type: string
          pattern: "(.*@.*\\..*)?"
      required:
        - first_name
        - last_name
      example:
        _id: "621ff9a936997df5baba746e"
        first_name: "Fox"
        last_name: "Mulder"
        email: "mulder@xfiles.com"
        phone: "+49 721 553181"
  securitySchemes: {}